requirements.txt
google-generativeai>=0.7.2
pypdf>=4.2.0
python-docx>=1.1.2
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.73.0
firebase-admin>=6.5.0
